Optimise list (non-)membership checks in GPC (#1736)
This PR suggests a way of more efficiently handling list
(non-)membership checks in GPCs. The main change is that the list
membership config keeps track of lists by comparison value identifier,
which makes more sense because ultimately we should only be doing at
most one list membership check per value rather than duplicating
indices. This PR thus drops the 'maxLists' count on circuits to the bare
minimum, but 'maxListElements' still falls back to the largest input
list size, which might not be indicative of the list size on the circuit
level.

I am pondering over whether it is worth keeping the current setup of
allowing both `isMemberOf` and `isNotMemberOf` fields or only allowing
one of them, as well as restricting these fields to naming a single list
rather than multiple lists. The burden of list combination logic would
then fall on the user of the library, but this would result in no real
loss of generality anyway and would get rid of the asymmetry we
currently have where list unions have to be computed manually.

Marking this as a draft for now. Comments welcome though!