Detailed docs and links in READMEs for POD+GPC (#1742)
This adds detailed docs to the README files of all POD/GPC related
packages. Most of the text content is borrowed from the developer page
in Notion. The bulk of the changes is creating a set of standard icons
and links so that each package links to all of its resources, and also
links to related packages.

Note that the README contents also shows upon the TypeDoc site at the
top of the package. That's good because it means no need to duplicate
contents in class docs. It's also why the header icons link to both
TypeDoc and GitHub, since we don't know which one the reader may be
coming from.