Add eddsa-pubkey type to POD (#1741)
Towards
https://linear.app/0xparc-pcd/issue/0XP-891/gpc-constraints-on-signing-keys

This PR adds an EdDSA public key type to the POD library. Such values
are tagged with `eddsa-pubkey` and encoded as hex strings in compressed
format as is done elsewhere. The simplified JSON serialisation adds the
prefix `pk:` to distinguish them from strings, which would either be
prefixed with `string:` or not follow the public key pattern.