Fixes broken doc site build (#1749)
The docs site is failing to build as, due to a type change, code in the
perftest package fails typechecking.

This was not causing a CI failure, because it doesn't get type-checked
in the `build` phase (it has no `build` command) nor in the `lint`
phase. I've added `tsc --noEmit` to the `lint` command (similar to what
we do in `passport-client`) to ensure type-checking of this package
during CI.