Add "pod_" prefix to types in raw value encoding (#1764)
Resolves
https://linear.app/0xparc-pcd/issue/0XP-928/pod-value-type-prefixes-collide-with-urls

Using `${type}:value` as a prefix, and checking for unknown types, has
the unfortunate property of colliding with URLs, because "https:" isn't
a known POD type. This is going to be error-prone for devs using
simplified JSON encoding.

I thought about alternate special characters, but I think the most
reliable way to ensure people don't accidentally run into a POD-related
field when they're not doing POD-related stuff is to add a `pod_`
prefix, like we do with PODNames. The format is now `pod_${type}:value`.

---------

Co-authored-by: Ahmad Afuni <root@ahmadafuni.com>