Use unpadded base64 instead of base64url (#1765)
The `base64url` encoding turns out not to be supported by the polyfilled
Buffer implementation we use in a browser, or even the one in the
"buffer" package, so I'm getting rid of it. Instead I'm manually
stripping padding off of base64, which is what I cared about from
base64url anyway. Note that decoding always accepts unpadded input
regardless.

Tested all of the POD operations in consumer-client this time to be sure
encoding and decoding work as expected.

Flagging @robknight to review since he reviewed the original version.
CC @ichub since this fixes the issue you ran into in passport-client.
CC @ax0 in case this impacts your ongoing work. Note that our default
test private key conveniently contains none of the characters which
differ between base64 and base64url, so copies of it around our codebase
don't need to be updated.