Remove dependency of dev server on gen-test-artifacts (#1771)
Waiting for circuit generation to start the dev server is now
unnecessary for most use cases, and it wastes people's time every time
they change branches or pull updated code. In order to use these
artifacts at all, you need to be doing active circuit development, and
you need to change the `GPC_ARTIFACTS_CONFIG_OVERRIDE` variable in your
server .env file(s) to enable this.

This is a quick-and dirty fix, removing the automatic dependency on
generation. I also documented the env vars and added a note about the
need for manual generation. I don't expect anyone to notice this note,
but it's a stopgap until I can come up with a better solution.

Attn: @ax0 who's the other person likely to be affected by the need to
generate manually currently.