one click email send via podbox (#1784)
https://github.com/proofcarryingdata/zupass/assets/2636237/e44929a5-8cfc-4bee-880d-655ac7bf9d9a

#### summary
add new button to podbox that is only useable on the edge esmeralda
pipeline that when clicked queues an email for each unique ticket holder
email that 1) hasn't already been checked in and 2) hasn't been sent an
email yet, containing a 1-click sign in link.

the workflow going forward will be to click this button once per day. if
this becomes annoying we can automate further with a cron job.

by default this feature will *not* be enabled in staging or local. It's
also only enabled for the esmaralda pipeline:

![Screenshot 2024-06-13 at 1 04
48 PM](https://github.com/proofcarryingdata/zupass/assets/2636237/d42a74c5-f0d8-47a7-8117-59b57900a7fc)

#### todo before merging
import all email addresses to whom one of these emails has already been
sent via our manual workflow into the prod database, so that they don't
get duplicate emails.

#### example
I've tested this locally, and was able to get a one-click-login email
that successfully logged me into zupass.

![Screenshot 2024-06-13 at 12 53
26 PM](https://github.com/proofcarryingdata/zupass/assets/2636237/159f24be-0c9e-4bdf-b978-113844e355db)