Turborepo 2.0/watch mode (#1794)
Closes
https://linear.app/0xparc-pcd/issue/0XP-832/investigate-turborepo-watch-mode

This PR upgrades Turborepo to 2.0.4 and uses Turborepo's built-in watch
mode to run apps. This replaces the previous custom script.

The `dev` task is defined as depending on `^build`, meaning that the
`dev` task for an app like `passport-server` is dependent on `build`
having run for its dependency tree.

The command `turbo watch dev --filter=./apps/*` will run `dev` for all
apps. These are persistent tasks and will not be restarted by Turborepo.
If any of the dependencies for these tasks change, then those
dependencies will be rebuilt. The `dev` processes for each app will
notice a change in their dependencies and will restart themselves when
this occurs.

This produces much the same effect as the custom script, but with less
complexity.

Turborepo introduces a new UI for running parallel tasks, in which the
tasks are listed in a left-hand pane, with their outputs on the right.
By navigating between the tasks, you can see the outputs for each. This
can be disabled by setting `TURBO_UI=0`, e.g. `TURBO_UI=0 yarn dev`.

There are some small tweaks to the `turbo.json` files for version 2.0,
mostly due to `pipeline` being renamed to `tasks` and the `.env` file
being managed by `inputs` rather than a separate configuration option.