Add virtual entries to GPC (#1783)
This PR adds virtual entry support to the GPC layer, which amounts to
exposing the signer's public key as the entry `$signerPublicKey`.

Specifying this entry is optional and defaults to *not* revealing the
signer's public key. I'm contemplating changing this to revealing the
public key by default to ensure that users don't fall into the trap of
forgetting to constrain the public key.

---------

Co-authored-by: Andrew Twyman <artwyman@users.noreply.github.com>