Editing of CSV data in Podbox UI (#1816)
Closes
https://linear.app/0xparc-pcd/issue/0XP-982/better-ui-for-csv-pipeline

During both creation and editing of a CSV pipeline, it's now possible to
view and edit the CSV data using the spreadsheet view.

Editing via the spreadsheet view is equivalent to editing the text in
configuration view. It does not automatically save the changes to the
back-end, in the same way that editing the configuration view does not
auto-save either. The user can switch between the two views, making
edits in both, before deciding to save them. If the user edits the
configuration so that it is not valid JSON, the CSV spreadsheet view
will refuse to load. If the CSV content is edited so that it is no
longer a valid CSV (e.g. it has the wrong number of columns) then the
CSV parser will also reject it, with an error message.