Add bounds checks to GPCircuits (#1814)
Towards
https://linear.app/0xparc-pcd/issue/0XP-951/gpc-lowerboundupperbound-proofs
Depends on
[proofcarryingdata/snark-artifacts#6](https://github.com/proofcarryingdata/snark-artifacts/pull/6)

This PR introduces bounds/range checks on integer entry values in the
ProtoPODGPC. The approach taken here is the addition of a new module and
a template parameter specifying the number of checks.

Summary of changes:
- Addition of `BoundsCheckModule` on the Circom side as well as
TypeScript tests.
- Corresponding changes to ProtoPODGPC including a new circuit parameter
to indicate the number of bounds checks.
- Dummy data in GPC layer (for now).

Range checks for the inputs to the module have been kept to a minimum in
line with the comments on the Linear ticket: The bounds themselves are
public and so would be checked elsewhere, while the range check itself
guarantees that the value being checked lies in the appropriate range
under the assumption that the bounds do.

As mentioned above, bounds checks are optional and the number of
required checks is specified as a template parameter. However, the
overhead isn't too great so that it might make more sense to absorb it
into the entry module with a corresponding flag after all. Thoughts?

Note that the CI will fail due to the artifact change.

---------

Co-authored-by: Andrew Twyman <artwyman@users.noreply.github.com>