Use Base64 for private keys in example (#1822)
This gives the GPCPCD display card the same improvements as the PODPCD:
- Scrolling text boxes
- Prettier button for verification
- Verification error display
- Link to docs in description
- General text cleanup

Note that most users won't ever see this display, which only appears if
a proof PCD is added to your Zupass. It may as well be consistent and
nice, though, if/when it comes up. The prove screen is unchanged, and
already looks fine.