Basic Semaphore groups for CSV pipelines (#1824)
Adds basic Semaphore group support to CSV pipelines.

It is not possible to configure multiple groups. Instead, a single
setting called `semaphoreGroupName` is provided. If unset, there are no
Semaphore groups. If set, a single group will be created which includes
all of the consumers of the pipeline. Consumers will only be members of
the group after they have fetched their tickets at least once.

---------

Co-authored-by: Richard Liu <richard.y.liu@berkeley.edu>