Fix build output to solve ESM import issue (#1842)
**See later comments for revised approach.**

According to the ESM standard, ESM imports are meant to use a
fully-qualified filename, including the .js extension, when importing a
JS module, although some bundlers use import resolution strategies which
allow the extension to be omitted. When using Deno or other
standards-enforcing tools, an import without an extension will fail.

`tsc` will not add these extensions during transpilation. However, we
can add file extensions to the imports in our TypeScript source, and
these will be preserved when transpiled to JavaScript. I've added .js
extensions to all imports in our `packages` directory, and an `eslint`
rule to ensure that we continue to add them to all future imports.

The only downside is that the `.js` extensions look a bit ugly. We might
be able to find another solution, but no other option seems obvious
right now.

In the course of implementing this, I encountered some out-of-memory
errors in GitHub Actions during the `lint` job. I limited concurrency of
`lint` tasks to 4, which seems to have fixed it.