New POD pipeline (#1830)
Closes https://linear.app/0xparc-pcd/issue/0XP-1023/tabular-input
Closes
https://linear.app/0xparc-pcd/issue/0XP-1026/specify-pod-entries-for-output
Closes https://linear.app/0xparc-pcd/issue/0XP-1030/feed-based-issuance
Closes
https://linear.app/0xparc-pcd/issue/0XP-1045/non-matching-issuance
Closes
https://linear.app/0xparc-pcd/issue/0XP-1035/readwrite-access-for-json-config-restricted-to-admins-only

This PR introduces a new `PODPipeline` (this breaks the convention of
naming the pipeline after the data source, but the purpose of this
pipeline is to issue PODs; I'd be happy to take suggestions for a
different name).

The configuration options are fairly simple and mostly focused on the
mapping between input data and output PODs. Input data is provided in
tabular format and is stored as a CSV file, with a configurable set of
columns which must match those in the CSV data. These columns have types
(e.g. string, integer, date) and the input data is parsed into those
types; data which fails to parse is invalid.

The typed input data is then mapped to POD entries, with special
conversion logic for certain types (e.g. a `UUID` input is mapped to an
`int` `PODValue` by using `uuidToBigInt()`). Not all input types can be
converted into `PODValue`s, and the UI will prevent the selection of an
impossible conversion.

This is very much step 1 of making PODBox self-serve for "smart
non-technical users". The UI is meant to be simple enough to be easy,
but comprehensive enough that it should never be necessary to edit the
JSON configuration manually:

<img width="908" alt="image"
src="https://github.com/user-attachments/assets/0c9809fe-8db9-4bee-9031-03071672f575">

<img width="920" alt="image"
src="https://github.com/user-attachments/assets/5ec5de42-5001-4de8-92c3-eaf2493f7830">

# Usage

To begin, create a new pipeline of type "POD". When creating the
pipeline, you should by default see a table containing three columns.
These columns can be renamed or deleted, and you can add additional
columns. Each column has an icon indicating its type:
- String
- Date
- Integer (caveat: I may need to add some limitations on the size of the
data here to fit the POD int/cryptographic distinction)
- UUID
- Boolean

Some of these columns have custom UI, to assist in entering valid data.

Once created, it is still possible to edit the structure of the table,
adding, deleting or renaming input columns.

It is also possible to configure the Output, the PODs that will be
generated from the data table.

This is a pretty simple mapping of input data to outputs, using simple
type conversions. In some cases, multiple conversions are possible: a
boolean could be converted to the `int` 1n, or the `string` "true".
UUIDs can be converted to `string`s or `cryptographic` numbers, but not
to `int`s. Invalid conversions are not selectable in the UI.

The final important piece is the "Match" filter for outputs. By default,
no filter is selected, meaning that all PODs generated from the input
data will be issued to any user who subscribes. To restrict issuance,
you must select either "Email" or "Semaphore ID" as the matching field,
and then select the POD entry which contains either the email or
Semaphore ID value. With this selected, PODs will only be issued to
those users who have a matching credential.

# Scope

This covers the bare bones of self-serve POD issuance. Out of scope for
this PR are some useful things that we will need in future:
- Support for POD entries containing the EdDSA signing key (will be
necessary for POD tickets)
- Pre-set POD "types" with mandatory entries
- Direct upload of a CSV file
- Pipeline-or-user-specific signing keys

All of the above have Linear issues. Further feature suggestions would
be appreciated and should be added to Linear.

# Comment checklist

- [x] Separate package for shared PODBox client/server code
- [x] Configured values of types other than string
- [x] Separate "cryptographic" input type
- [x] ~~Use PODContent rather than PODEntries in some places~~