Ignore .tsbuildinfo files at the package level, not root level (#1852)
Closes
https://linear.app/0xparc-pcd/issue/0XP-1106/use-gitignore-files-in-individual-packages-to-ignore-tsbuildinfo-files

tldr this fixes an annoying problem when switching branches and running
`git clean`, which deletes orphaned build outputs but not orphaned
`.tsbuildinfo` files, which causes builds to fail when returning to the
branch with the additional package. More detailed explanation below.

---

`.tsbuildinfo` files are created by TypeScript during compilation. They
are incremental build artifacts which TypeScript uses to speed up
compilation - in particular, TypeScript will skip building a package if
the `.tsbuildinfo` file is newer than the source files.

Currently we avoid checking these files into the repo by referencing
them in the `.gitignore` at the root of the project. But this causes a
problem when switching branches, because the build outputs for each
package are ignored using the _package-level_ `.gitignore` file.

Let's say I'm on a branch called `new-package`, and I have created a new
package and built it. Then I switch to `main`. git will remove the files
from my new package which have been checked into the repo (the source
files, the package.json, and the .gitignore file, among others). But it
will not remove any files that are not checked in. So, my package
directory will remain, and it will contain some `.tsbuildinfo` files and
a `dist` directory with JavaScript outputs. When I run `git status` I
will see that there's an unstaged directory, and I will think "aha, I
need to run `git clean -f -d` to get rid of this", which is correct.

`git clean` will remove all _un-ignored_ files in that directory. Since
the `dist` directory, containing the build output, is no longer being
ignored by the package-level `.gitignore` file, it will be removed.
However, the `.tsbuildinfo` files are ignored by the root-level
`.gitignore`, so they will remain.

After I return to my `new-package` branch, I will have the
`.tsbuildinfo` files but not the build output from `dist`. If I try to
run a build, `tsc` will think it can skip building the package, which is
incorrect. I will get a build failure, and will have to go to the
package directory, remove the `.tsbuildinfo` files, and then build the
package again. Now it will work.

This is happening because the `.tsbuildinfo` files are being ignored at
the package root rather than in the package-local .`gitignore` file,
which means that the `.tsbuildinfo` files don't get deleted along with
the `dist` directory when doing a `git clean` after switching branches.
This PR fixes that by removing the `*.tsbuildinfo` reference from the
root `.gitignore`, and adding it to each of the individual package-level
`.gitignore` files.