Prototype zapp API (#1854)
Basic code organization:

The API is implemented in two main parts:
1. The Zupass client, in the `zupass-client` package
2. The Zapp server, implemented inside `passport-client`

The client is responsible for loading an iframe, sending an initial
message to it, and then creating a `Client` object which the library
users can use to interact with the embedded Zupass iframe.

The initial bootstrapping is done using messages described in
`protocol.ts`. This involves "window messages" which are used to
initiate the connection, then "RPC messages" which are sent over a
dedicated `MessageChannel` port. The RPC messages are used to invoke API
calls and to send either successful results or errors back. Each RPC
invocation has a serial number which is used to correlate messages
containing results or errors with the invocations.

On the client side, a JavaScript `Proxy` object is used to present an
API object. So when client-side applications call `z.fs.list` or
`z.gpc.prove`, those calls go to the proxy object which then works out
what kind of message to send. Since JavaScript lacks reflection, we use
a Zod schema (in `schema.ts`) as a map of the API endpoints. The API
endpoints are described canonically in `api.ts`.

The client package also has some logic for managing a modal dialog,
inside which the iframe can be displayed.

On the "server" side, a similar mapping occurs. Here, in `ZappServer.ts`
we have the API implementation. This is a set of classes for the
different API sections (`fs`, `gpc`, `identity`, and so on). These
implement the API described in `api.ts` in the `zupass-client` package.
The `@safeInput` decorator is used to ensure that the parameters to
these methods are of the correct type; arguably this feature belongs in
the lower-level protocol.

The `useZappServer` hook is used to actually instantiate the server
component inside `passport-client`. This needs some refactoring to
support connections that go from Zupass to an embedded Zapp, rather than
the other way around.

Finally there is the `EmbeddedScreen` and its children. When Zupass is
embedded inside the client app, it navigated to this screen via the
`/embedded` route. Application state is then used to change the contents
of the screen, e.g. to show a prove screen or an add subscription
screen.

---------

Co-authored-by: Richard Liu <richard.y.liu@berkeley.edu>