Prototype query API (#1858)
A prototype query API for querying PODs.

Most of the query logic is implemented by a new library called
`podspec`, which allows for the description of a POD, its entries,
signing key, and certain constraints on those attributes.

These can be used to define a "shape" which can be used for
pattern-matching against a single POD or POD collection, and there is a
simple query API implementation which does this by pattern-matching
against all of the user's `PODPCD`s.

See the `podspec` README file for details of how this works, with
examples given in the tests.