Fix GPC owner module compilation (#1865)
This PR addresses a GPC compiler bug where the owner nullifier hash is
not properly revealed when both an owner ID entry and external nullifier
are specified in the case where the owner identity commitment entry has
index 0. This may be checked by generating a proof for a simple
configuration such as
```js
{
    pods: {
        pod0: {
            entries: {
                owner: { isRevealed: false, isOwnerID: true }
            }
        }
    }
}
```
with appropriate proof inputs and inspecting the revealed nullifier
hash, e.g. by commenting out the `attack` entry of the `weapon` POD in
`simpleProofConfig` in `examples/pod-gpc-example/src/gpcExample.ts` and
checking `revealedClaims.nullifierHash`, whose value is then
`BABY_JUB_NEGATIVE_ONE`. Appropriate tests have been added to the `gpc`
package to check for this.