Replace call to "assert.equal" with manual check and exception (#1871)
Closes
https://linear.app/0xparc-pcd/issue/0XP-1142/remove-assert-dependency

The `assert` module is built in to NodeJS, but not in browser contexts.
It can be polyfilled, but the polyfill is ~80KB and can't be reduced by
tree-shaking. It also requires consumers to figure out why `assert`
isn't working and how to install the relevant polyfill, which is
annoying for them.

We're only using it for `assert.equal`, and only in one place (in
non-test code).

According to the docs, `assert.equal` is an alias for
`[assert.strictEqual](https://nodejs.org/api/assert.html#assertstrictequalactual-expected-message)`,
which tests equality according to the criteria used by `Object.is`. We
can replace our `assert.equal` call with a check using `Object.is`,
throwing our own exception if the check fails, and avoid the need for
the polyfill. I've approximately replicated the error message that
`assert.strictEqual` would show.

Janabel ran into this issue recently, and I just encountered it myself,
so I presume that this is reasonably common.