Insert and delete PODs, with insertion demo in test-client (#1863)
Stacks on top of #1861, which should probably be reviewed first.

This demonstrates adding and deleting PODs. Since PODs are immutable,
there isn't really an "update" operation. To replace a POD with a new
version, simply delete the old one and insert a new one.

Most of the code here is the `test-client` UI, which allows you to
create a brand-new POD interactively. The actual API implementation is
very simple.

---------

Co-authored-by: Richard Liu <richardliuwa@gmail.com>