Fix SnarkJS type dependencies (#1877)
In several packages we depend on the `snarkjs` library. However, in each
of these cases we do not include a devDependency on `@types/snarkjs`.
This means that when our packages referencing SnarkJS types are used by
third parties, they lack type information.

I noticed this because the `Groth16Proof` type is used in the return
value for `gpcProve`, and was being interpreted by TypeScript as `any`.

This did not show up for us inside the monorepo, because the root
package.json has a direct dependency on `@types/snarkjs`, and `yarn`
does not ensure that each individual package which depends on those
types has its own dependency entry. I've fixed this by removing the
dependency from the root package, and adding individual dependencies to
each package which also depends on `snarkjs`.