Add missing lodash dependency to gpc package (#1887)
Adds a missing `lodash` dependency to `@pcd/gpc`, which I noticed when
installing `@pcd/gpc` in a fresh repo.

Annoying that the `no-extraneous-dependencies` ESLint rule didn't catch
it, I assume that the presence of the `@types/lodash` package was
satisfying the requirement that the module be declared, even if the
implementation was missing.