Use Parcnet client libraries to implement Zapp API (#1888)
This brings the Zupass implementation in line with the latest Parcnet
API version, with support for POD signing.

POD signing can also be configured to be restricted to certain origins -
see `env.local.example` in `passport-client`.

The embedded GPC proof screen is experimental and partially complete,
lacking support for `equalsEntry` and `notEqualsEntry` checks. A
dedicated proof screen for tickets will follow shortly.

I've also updated the `test-client` app to support all of the latest
APIs.

The `zupass-client` package has been removed as this is now superseded
by `@parcnet-js/app-connector`, and the code in `passport-client` has
been simplified as it now relies on `@parcnet-js/client-helpers` to
handle iframe messaging, validation, etc.

---------

Co-authored-by: Richard Liu <richardliuwa@gmail.com>