Basic ticket proofs using Z API (#1891)
The easiest way to test this is simply to run the test Zapp and see the
code that is used. It's a thin wrapper over the GPC proving mechanism at
the moment, and the prove screen needs more work, but this gives a very
simple API for developers to authenticate tickets.