Authenticate embedded Zupass using localStorage via StorageAccessAPI (#1864)
Closes https://linear.app/0xparc-pcd/issue/0XP-1111/log-in-via-cookie
Closes
https://linear.app/0xparc-pcd/issue/0XP-1139/use-requeststorageaccess-to-access-local-storage

In most browsers, local storage is "partitioned" by top-level origin.
This means that if I visit zupass.org, I can create cookies, local
storage entries, and so on, and these will be available to other tabs or
windows pointing to zupass.org (in the same browser).

However, iframes do not work like this. They will not, by default, share
local storage with other instances. To date, we simply allow the user to
log in via the iframe, which downloads their storage. However, requiring
the user to enter their username and password adds some friction.

~The only universally-supported way around this is to call
`requestStorageAccess`, which grants access to first-party _cookies_ in
a third-party (iframe) context. In this PR, Zupass sets a cookie
containing the user's encryption key when they log in. When Zupass is
loaded in an iframe, storage access is requested and, if approved, is
used to read the encryption key from the cookie and log the user in.~

Using cookies to transfer information to iframes is not sufficiently
secure; see comments below.

However, we can use a Chrome-specific API which allows iframes to access
first-party local storage, with user permission. This allows us to
retrieve the information necessary to boostrap a user session. This will
only work in Chrome and derivatives like Arc, but since this is around
66% of browsers by market share, it seems worth doing.

This reduces the interaction required when using Zupass in an iframe
from entering an email and password and clicking a button to simply
clicking a button and clicking a further confirmation to grant storage
access.

In cases where we detect that the user is running Chrome, instead of
showing a request for email and password, we show this:

<img width="517" alt="image"
src="https://github.com/user-attachments/assets/900768fd-d810-4664-ad13-a6d50975d5bd">

Clicking the button will request storage access:
<img width="941" alt="image"
src="https://github.com/user-attachments/assets/c47d8aba-1fb3-4655-b5d5-e227c5c5a1b2">


If the user approves, and the local storage contains a valid encryption
key, then we hash the encryption key and request the encrypted storage
from the server. If this decrypts successfully then we use the storage
to log the user in.

On subsequent visits, the embedded Zupass will remember the user and
will not require clicking a button or logging in.