Bump `@zk-kit/baby-jubjub` and `@zk-kit/lean-imt` (#1893)
In #1883 I bumped the version of `@zk-kit/eddsa-poseidon` which
`@pcd/pod` depends on. However, the new version of
`@zk-kit/eddsa-poseidon` depends on a higher version of `@zk-kit/utils`
than `@zk-kit/baby-jubjub`, and now both of those versions get included
in the bundle. This seems to be because `@zk-kit` specifies their
dependencies with fixed version numbers (`1.2.0` rather than `^1.2.0` or
`~1.2.0`) and so it's not possible to satisfy both dependency
requirements with the same version even if the difference between
dependency requirements is very small.

Secondly, we also do the same thing with our dependencies on
`@zk-kit/*`, specifying the exact version numbers rather than just the
major version number (by prefixing the version with `^`) or the minor
version number (by prefixing with `~`).

We also have a dependency on `@zk-kit/lean-imt` which pulls in a _third_
version of `@zk-kit/utils`, because we depend on
`@zk-kit/lean-imt@2.0.1`, which depends on `@zk-kit/utils`.

If we think that the `@zk-kit` project is using semantic versioning
correctly, we should consider changing our dependencies to use the `^`
prefix, which would allow newer versions of the `@zk-kit` packages to be
used in dependency resolution, and reduce the likelihood of having
multiple versions of packages being included in JavaScript bundles in
order to satisfy overly restrictive dependency requirements.