add `gpcCheckProvable` as a "dry run" of `gpcProve` (#1895)
This splits out the check, canonicalize, and circuit-requirements parts
of `gpcProve` so that it's possible to check the validity of args
without paying the cost of compiling circuit inputs or generating a
proof. The new `gpcCheckProvable` function is similar in concept (and
identical in return type) to `gpcBindConfig`, but differs in that it
takes into account specific proof inputs, not only the circuit
configuration.

The new function is directly called by `gpcProve` so there should be
little chance of it falling out of sync. I enhanced the tests to ensure
the same errors are thrown by both functions, to ensure we don't end up
with (known) cases where the compile phase is throwing errors which
belong in the check phase.

CC @robknight who may have use for this filtering or querying for the
ZAPI