Embedded Zapps in Zupass (#1899)
This adds functionality for Zapps to be embedded in Zupass.

This can be enabled by configuring the `EMBEDDED_ZAPPS` environment
variable for `passport-client`. This is a stringified JavaScript object,
of the form:
```json
{ "Zapp name": "http://zapp-url.com" }
```

A folder will appear in Zupass using the given name, and clicking on the
folder will take the user to the Zapp.

Zapps will require some minimal changes to be able to support embedding,
which are shown in the `test-zapp` example code.