Iframe login via popup (#1905)
This allows the user to log in via a popup window opened from within the
Zupass embedded iframe.

On visiting a Zapp for the first time, the user is asked to connect to
Zupass by clicking a button, which causes a popup window to open.

If the user is already logged in to Zupass in another tab, then the
popup window will pick up the authentication state from that other
session, and forward the user's sync key back to the iframe, which will
use it to log in.

If the user is not already logged in to Zupass in another tab, then they
can provide their email and password in the popup window, and after
authenticating the sync key will be sent to the iframe.

Three failure cases are covered:

1) If the browser blocks popups, a message is shown to the user advising
them to change their browser settings and try again
2) If the user closes the popup window before authentication completes,
they are informed of this and asked to try again
3) If logging in using the sync key fails, the user is asked to try
again