UnknownPCD (#1881)
UnknownPCD acts as a wrapper for a PCD which fails to deserialize. This
allows the rest of the PCD collection to remain usable. It also allows
the wrapped "bad" PCD to be preserved and re-serialized in its original
form so that a later version of Zupass can make another attempt to
deserialize it. The intent is that this creates a smoother, and more
recoverable scenario in the case of bugs or incompatibilities. E.g.
GPCPCD's format is not currently stable, so anyone who saves one of them
into their Zupass may run into errors which UnknownPCD is intended to
address.

This draft is not yet complete, but I'd like to get some feedback on the
design approach before I finish it. What's completed:

- UnknownPCD package behavior
- Wrapping functions to create UnknownPCD, and derive a PCD ID
- Configurable fallback function in PCDCollection deserialization
- Use of that fallback function in Passport-Client when reading from
local and E2EE storage

There are TODOs in the code indicating the parts which are incomplete,
most notably the PCD UI. I haven't tested this in a real Zupass client
yet, but the unit tests pass.

---------

Co-authored-by: Ivan Chub <ichub@users.noreply.github.com>