Vertical scroll for POD & GPC PCDs (#1908)
Add vertical scrolling to POD-PCD and GPC-PCD
Closes
https://linear.app/0xparc-pcd/issue/0XP-1087/vertical-scrolling-for-podgpc-display-text

Note this PR is dependent on #1881. This happened because my local test
server is broken without UnknownPCD. Which I could ignore, but I see
this is a reasonable test case to see how to get stacked PRs to work.
I've marked this as merging into the parent PR's branch, and I'll figure
out later how to update it when the parent PR merges.