Remove Strich scanner (#1912)
This turns out not to be very important any more, given that scanning
for check-in is no longer required. We don't seem to have kept up our
license in any case, as in production the SDK fails to initialize and we
get a 403 error from the Strich server! This means that we're already
failing over to `react-qr-reader`, and we might as well make that the
default (only) scanner option.

This will save some 587KiB from our bundle size.