Only depend on Semaphore identity v4, without requiring Semaphore core (#1923)
This is a small bit of tidying up. In order to import Semaphore identity
v4, we have been importing `@semaphore-protocol/core`, which happens to
also import and re-export `@semaphore-protocol/identity` v4.x - we can't
just import `@semaphore-protocol/identity` because we're already
importing version `3.15.2`.

However, we can simply alias `@semaphore-protocol/identity@^4.0.3` to a
different name, and import it that way, thereby avoiding the extra code
that comes with `@semaphore-protocol/core`.