Replace external constant which has side-effects with internal constant declaration that does not (#1914)
In `podTypes.ts` we import the `r` constant from `@zk-kit/baby-jubjub`.
This is declared
[here](https://github.com/privacy-scaling-explorations/zk-kit/blob/main/packages/baby-jubjub/src/baby-jubjub.ts)
in `@zk-kit/baby-jubjub`, in a file which contains a mix of constant
declarations and side-effecting imports. In particular, it imports and
instantiates the `F1Field` class, which means that it's also necessary
to import `@zk-kit/utils/f1-field`, and it also imports
`leBigIntToBuffer`, which depends on the Buffer polyfill.

I'm a bit disappointed that neither Rollup nor esbuild are capable of
detecting that we only want this single constant and not any of the
other things imported or instantiated in the same file that the constant
is declared in, but after testing various confgurations it seems that
they truly are not.

Therefore, to avoid consumers of `@pcd/pod` having to import these
external dependencies, we can re-declare the constant and verify that it
has the correct value using tests.